Move Google Sheets fetching to server
