+++ server.js
@@ -3,6 +3,7 @@ const fs = require('fs').promises;
 const path = require('path');
 const multer = require('multer');
 const cors = require('cors');
+const { google } = require('googleapis');
 
 const app = express();
 const port = process.env.PORT || 3001;
@@ -16,6 +17,15 @@ const upload = multer();
 
 // File paths
 const DATA_FILE = path.join(__dirname, 'data', 'events.json');
+
+// Google Sheets setup
+const sheets = google.sheets('v4');
+const auth = new google.auth.GoogleAuth({
+  keyFile:
+    process.env.GOOGLE_SERVICE_ACCOUNT_KEY ||
+    process.env.REACT_APP_GOOGLE_SERVICE_ACCOUNT_KEY,
+  scopes: ['https://www.googleapis.com/auth/spreadsheets.readonly'],
+});
 
 // Ensure data directory exists
 const ensureDataDir = async () => {
@@ -72,6 +82,33 @@ app.post('/api/sync', upload.none(), async (req, res) => {
   } catch (error) {
     console.error('Error syncing data:', error);
     res.status(500).json({ error: 'Failed to sync data' });
+  }
+});
+
+app.get('/api/registrants', async (req, res) => {
+  const { spreadsheetId, range } = req.query;
+  if (!spreadsheetId || !range) {
+    return res.status(400).json({ error: 'Missing spreadsheetId or range' });
+  }
+
+  try {
+    const authClient = await auth.getClient();
+    const response = await sheets.spreadsheets.values.get({
+      auth: authClient,
+      spreadsheetId,
+      range,
+    });
+
+    const registrants = (response.data.values || []).map(row => ({
+      name: row[0],
+      email: row[1],
+      phone: row[2],
+    }));
+
+    res.json(registrants);
+  } catch (error) {
+    console.error('Error fetching data from Google Sheets:', error);
+    res.status(500).json({ error: 'Failed to fetch data from Google Sheets' });
   }
 });
